Верстка проекта. 2 Спринт.
