+++ src/components/modal-overlay/modal-overlay.tsx
@@ -1,35 +1,12 @@
-import React, {useEffect, useRef} from "react";
+import React, {ForwardedRef, RefObject} from "react";
 import styles from "./modal-overlay.module.css"
 
-type Props = { onClose: () => void; children: React.ReactNode;};
+type Props = { children: React.ReactNode; };
 
-export default function ModalOverlay(props: Props) {
-    const {onClose, children} = props;
-    const rootRef = useRef<HTMLDivElement>(null);
+export const ModalOverlay = React.forwardRef((props: Props, ref: ForwardedRef<HTMLDivElement>) => {
+  const {children} = props;
 
-    useEffect(() => {
-        const handleWrapperClick = (event: MouseEvent) => {
-            const { target } = event;
-
-            if (target instanceof Node && rootRef.current === target) {
-                onClose?.();
-            }
-        };
-        const handleEscapePress = (event: KeyboardEvent) => {
-            if (event.key === 'Escape') {
-                onClose?.();
-            }
-        };
-
-        window.addEventListener('click', handleWrapperClick);
-        window.addEventListener('keydown', handleEscapePress);
-
-        return () => {
-            window.removeEventListener('click', handleWrapperClick);
-            window.removeEventListener('keydown', handleEscapePress);
-        };
-    }, [onClose]);
-    return <div className={styles.modalOverlay} ref={rootRef}>
-        {children}
-    </div>
-}
+  return <div className={styles.modalOverlay} ref={ref}>
+    {children}
+  </div>
+})
