+++ src/components/order-details/order-details.module.css
@@ -2,4 +2,10 @@
     display: flex;
     flex-direction: column;
     justify-content: center;
+    align-items: center;
+}
+
+.image {
+    width: 120px;
+    height: 120px;
 }
